% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_variab.R
\name{al_variab}
\alias{al_variab}
\title{resample age-length data to implement growth variability}
\usage{
al_variab(age_dat)
}
\arguments{
\item{age_dat}{age specimen data}
}
\description{
resample age-length data to implement growth variability
}
