% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apop.R
\name{apop}
\alias{apop}
\title{Compute population at age}
\usage{
apop(lpop, agedat, lngs, by_strata = FALSE, global = FALSE)
}
\arguments{
\item{lpop}{population at length dataframe created from surveyISS::lpop_gap()}

\item{agedat}{age-length specimen input dataframe}

\item{lngs}{all combinations of possible lengths and ages}

\item{by_strata}{Boolean. Are the pop'n data structed by strata or summed to region level? (default = FALSE)}

\item{global}{Boolean. Fill in missing length bins with global age-length key? (default = FALSE)}
}
\value{
dataframe of population at age by year and species, also includes mean length-at-age and sd in length-at-age
}
\description{
Function to expand age-length specimen data to population at age aligned with gapindex package.
}
