% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apop_caal.R
\name{apop_caal}
\alias{apop_caal}
\title{conditional age-at-length}
\usage{
apop_caal(agedat, sex_spec = TRUE)
}
\arguments{
\item{agedat}{age dataframe}

\item{sex_spec}{determine whether computing for sex-specific or total comps}
}
\description{
conditional age-at-length
}
