% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apop.R
\name{apop_gap}
\alias{apop_gap}
\title{population at age following computations in gapindex package}
\usage{
apop_gap(lpop, agedat, lngs, by_strata = FALSE, global = FALSE)
}
\arguments{
\item{lpop}{length population data}

\item{agedat}{age dataframe}

\item{lngs}{all combinations of possible lengths and ages}

\item{by_strata}{are the length pop'n by strata or summed to region level (default = FALSE)}

\item{global}{fills in missing length bins with global alk (default = FALSE)}
}
\description{
population at age following computations in gapindex package
}
