% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{bias_age}
\alias{bias_age}
\title{calculate bias in bootstrapped age comp}
\usage{
bias_age(r_age, oga)
}
\arguments{
\item{r_age}{iterated age pop'n}

\item{oga}{original age comps without any sampling}
}
\description{
calculate bias in bootstrapped age comp
}
