% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{bias_caal}
\alias{bias_caal}
\title{calculate bias in bootstrapped conditional age-at-length}
\usage{
bias_caal(r_caal, ogcaal)
}
\arguments{
\item{r_caal}{iterated conditional age-at-length}

\item{ogcaal}{original conditional age-at-length without any sampling}
}
\description{
Computes the mean bias in bootstrapped conditional age-at-length compared to original unsampled
conditional age-at-length.
}
