% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_age.R
\name{boot_age}
\alias{boot_age}
\title{Resample age data w/replacement.}
\usage{
boot_age(age_dat)
}
\arguments{
\item{age_dat}{age-length specimen input dataframe}
}
\value{
dataframe of resampled sex-length-age by year, species, and haul
}
\description{
Resample age data w/replacement.
}
