% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_haul.R
\name{boot_haul}
\alias{boot_haul}
\title{Resample hauls w/replacement.}
\usage{
boot_haul(cpue_data)
}
\arguments{
\item{cpue_data}{catch-per-unit effort input dataframe}
}
\value{
dataframe of resampled hauls by year and species
}
\description{
Resample hauls w/replacement.
}
