% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_length.R
\name{boot_length}
\alias{boot_length}
\title{Resample length frequency data w/replacement.}
\usage{
boot_length(lfreq_un)
}
\arguments{
\item{lfreq_un}{expanded length frequency data (one row per sample)}
}
\value{
dataframe of resampled length-sex pairs by year, species, and haul
}
\description{
Resample length frequency data w/replacement.
}
