% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{comp_stats}
\alias{comp_stats}
\title{wrrapper function to compute statistics of bootstrap resampling}
\usage{
comp_stats(
  r_age,
  oga,
  r_length,
  ogl,
  specimen_data,
  lfreq_data,
  survey_region = NULL
)
}
\arguments{
\item{r_age}{list of abundance by age data}

\item{oga}{original abundance by age data (single list)}

\item{r_length}{list of abundance by length data}

\item{ogl}{original abundance by length data (single list)}

\item{specimen_data}{input dataframe}

\item{lfreq_data}{input dataframe}

\item{survey_region}{overall region, e.g., goa or ai (default = NULL)}
}
\description{
wrrapper function to compute statistics of bootstrap resampling
}
