% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_caal.R
\name{ess_caal}
\alias{ess_caal}
\title{calculate effective sample size for age comps}
\usage{
ess_caal(sim_data, og_data, sex_spec)
}
\arguments{
\item{sim_data}{list of abundance by age data}

\item{og_data}{original abundance by age data (single list)}

\item{sex_spec}{determine whether to do sex specific or total comps (default = TRUE)}
}
\description{
calculate effective sample size for age comps
}
