% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{grwth_stats}
\alias{grwth_stats}
\title{calculate mean and SD in length-at-age}
\usage{
grwth_stats(r_age, oga)
}
\arguments{
\item{r_age}{iterated population at age with mean length-at-age and sd in length-at-age output}

\item{oga}{original population at age without any sampling with mean length-at-age and sd in length-at-age output}
}
\description{
Calculate mean length-at-age and SD in length-at-age as aligned with gapindex package.
}
