% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{iss_length}
\alias{iss_length}
\title{calculate length comp input sample size and add nominal sample size and number of sampled hauls to output}
\usage{
iss_length(rss_length, lfreq_data, survey_region = NULL)
}
\arguments{
\item{rss_length}{iterated realized sample size}

\item{lfreq_data}{input dataframe}

\item{survey_region}{overall region, e.g., goa or ai (default = NULL)}
}
\description{
calculate length comp input sample size and add nominal sample size and number of sampled hauls to output
}
