% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpop.R
\name{lpop_gap}
\alias{lpop_gap}
\title{Compute population at length}
\usage{
lpop_gap(lfreq_un, cpue_data, by_strata = FALSE)
}
\arguments{
\item{lfreq_un}{expanded length frequency data (one row per sample) input dataframe}

\item{cpue_data}{catch-per-unit effort input dataframe}

\item{by_strata}{Boolean. Are the results to be given at strata level, or summed to region level (default = FALSE)}
}
\value{
dataframe of population at length by year and species
}
\description{
Function to expand length frequency data to population at length aligned with gapindex package.
}
