% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gap.R
\name{match_gap}
\alias{match_gap}
\title{compare with gap estimates of pop'n at age and length}
\usage{
match_gap(oga, ogl, gap_apop, gap_lpop, thresh = 0.01, region = NULL)
}
\arguments{
\item{oga}{surveyISS computed pop'n at age}

\item{ogl}{surveyISS computed pop'n at length}

\item{gap_apop}{gap computed pop'n at age}

\item{gap_lpop}{gap computed pop'n at length}

\item{thresh}{threshold for desired match between gap and surveyISS (set at 0.01 percent difference)}

\item{region}{region for which your comparing}
}
\description{
compare with gap estimates of pop'n at age and length
}
