% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gap_bs.R
\name{match_gap_bs}
\alias{match_gap_bs}
\title{compare with gap estimates of pop'n at age and length for bering sea}
\usage{
match_gap_bs(ogl, gap_lpop, thresh = 0.01, region = NULL, save = NULL)
}
\arguments{
\item{ogl}{surveyISS computed pop'n at length}

\item{gap_lpop}{gap computed pop'n at length}

\item{thresh}{threshold for desired match between gap and surveyISS (set at 0.01 percent difference)}

\item{region}{region for which your comparing}

\item{save}{name to save match output}
}
\description{
compare with gap estimates of pop'n at age and length for bering sea
}
