% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{read_test}
\alias{read_test}
\title{Age and growth lab reader-tester paired otolith data}
\format{
A data frame with 212158 observations and 4 variables:
\describe{
\item{species_code}{RACE species idenfication number}
\item{region}{survey region: AI, BS, GOA}
\item{age}{original reader age}
\item{test_age}{duplicated test age from different reader}
}
}
\usage{
read_test
}
\description{
A dataset containing duplicate reads of same otolith by different readers
}
\keyword{datasets}
