% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sampling.R
\name{red_len_samples}
\alias{red_len_samples}
\title{Reduce length frequency samples}
\usage{
red_len_samples(
  data,
  samples = NULL,
  grp = c("year", "species_code", "stratum", "hauljoin")
)
}
\arguments{
\item{data}{length sample dataframe}

\item{samples}{number of lengths to take (default = NULL)}

\item{grp}{grouping variable(s) (default is to reduce haul-level sampling)}
}
\description{
Function to test reductions in length sampling for whatever grouping is desired (default is reduction in haul-level sampling)
}
