% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{rss_caal}
\alias{rss_caal}
\title{Calculate conditional age-at-length realized sampled size (rss)}
\usage{
rss_caal(sim_data, og_data)
}
\arguments{
\item{sim_data}{list of replicated conditional age-at-length}

\item{og_data}{original conditional age-at-length}
}
\description{
Function to calculate rss for conditional age-at-length following McAllister and Ianelli 1997.
}
