% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{rss_length}
\alias{rss_length}
\title{calculate realized sample size for length comps}
\usage{
rss_length(sim_data, og_data)
}
\arguments{
\item{sim_data}{list of abundance by length data}

\item{og_data}{original abundance by length data (single list)}
}
\description{
calculate realized sample size for length comps
}
