% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_comps.R
\name{srvy_comps_ai_cmplx}
\alias{srvy_comps_ai_cmplx}
\title{primary survey expansion for age/length pop'n numbers function
 customized for ai complexes (e.g., blackspotted-rougheye rockfish)
  where length pop'n are estimated at species level, 
  but age pop'n are estimated at complex level}
\usage{
srvy_comps_ai_cmplx(
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  r_t,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  cmplx_code = NULL,
  use_gapindex = TRUE,
  by_strata = FALSE,
  global = FALSE
)
}
\arguments{
\item{lfreq_data}{length frequency data}

\item{specimen_data}{age-length specimen data}

\item{cpue_data}{abundance by length data}

\item{strata_data}{strata and associated area}

\item{r_t}{reader/tester ageing data}

\item{yrs}{age filter returns years >= (default = NULL)}

\item{bin}{length bin size (default = 1 cm)}

\item{boot_hauls}{switch for resampling hauls (default = FALSE)}

\item{boot_lengths}{switch for resampling lengths (default = FALSE)}

\item{boot_ages}{switch for resampling ages (default = FALSE)}

\item{al_var}{switch for including age-length variability (default = FALSE)}

\item{al_var_ann}{resample age-length annually or pooled across years}

\item{age_err}{switch for including ageing error (default = FALSE)}

\item{cmplx_code}{numeric value to replace the individual species codes with a complex code (default = NULL)}

\item{use_gapindex}{use functions derived from gapindex package (default = TRUE)}

\item{by_strata}{should length/age pop'n values be computed at stratum level in gap fcns (default = FALSE)}

\item{global}{fills in missing length bins with global alk in gap fcns (default = FALSE)}
}
\description{
primary survey expansion for age/length pop'n numbers function
 customized for ai complexes (e.g., blackspotted-rougheye rockfish)
  where length pop'n are estimated at species level, 
  but age pop'n are estimated at complex level
}
