% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_comps.R
\name{srvy_comps_caal}
\alias{srvy_comps_caal}
\title{primary survey condition age-at-length function}
\usage{
srvy_comps_caal(
  specimen_data,
  cpue_data,
  r_t,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE
)
}
\arguments{
\item{specimen_data}{age-length specimen data}

\item{cpue_data}{abundance by length data}

\item{r_t}{reader/tester ageing data}

\item{yrs}{returns years >= (default = NULL)}

\item{bin}{bin size for length data}

\item{boot_hauls}{switch for resampling hauls (default = FALSE)}

\item{boot_ages}{switch for resampling ages (default = FALSE)}

\item{al_var}{switch for including age-length variability (default = FALSE)}

\item{al_var_ann}{resample age-length annually or pooled across years}

\item{age_err}{switch for including ageing error (default = FALSE)}
}
\description{
primary survey condition age-at-length function
}
