% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_iss.R
\name{srvy_iss_ai_cmplx}
\alias{srvy_iss_ai_cmplx}
\title{Stock complex survey input sample size function}
\usage{
srvy_iss_ai_cmplx(
  iters = 1,
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  len_samples = NULL,
  age_samples = NULL,
  plus_len = NULL,
  plus_age = NULL,
  cmplx_code = 3005012,
  by_strata = FALSE,
  global = FALSE,
  cmplx = "bsre",
  region = "ai",
  save_interm = FALSE,
  save_stats = FALSE,
  save
)
}
\arguments{
\item{iters}{number of iterations (min of 500 recommended for full run)}

\item{lfreq_data}{length frequency input dataframe}

\item{specimen_data}{age-length specimen input dataframe}

\item{cpue_data}{catch-per-unit effort input dataframe}

\item{strata_data}{strata id and area size input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{bin}{bin size (default = 1 cm), also can use custom length bins through defining vector of upper length bin limits, i.e., c(5, 10, 20, 35, 60), plus length group will automatically be populated and denoted with the largest defined bin + 1 (i.e., 61 for provided example)}

\item{boot_hauls}{Boolean. Resample hauls w/replacement? (default = FALSE)}

\item{boot_lengths}{Boolean. Resample length frequency w/replacement? (default = FALSE)}

\item{boot_ages}{Boolean. Resample ages w/replacement? (default = FALSE)}

\item{al_var}{Boolean. Include age-length variability in resampled age data? (default = FALSE)}

\item{al_var_ann}{Boolean. Resample age-length variability annually or pooled across years? (default = FALSE)}

\item{age_err}{Boolean. Include ageing error in resampled age data? (default = FALSE)}

\item{len_samples}{If set at a value, tests reductions in haul-level length sampling. To test, set this value at some smaller level than current sampling rate, i.e., 25 (default = NULL)}

\item{age_samples}{If set at a value, tests reductions (and increases) in survey-level number of ages collected. To test, set at a proportion of ages collected, i.e., 0.8 or 1.2 (default = NULL)}

\item{plus_len}{If set at a value, computes length expansion with a plus-length group (default = FALSE)}

\item{plus_age}{If set at a value, computes age expansion with a plus-age group (default = FALSE)}

\item{cmplx_code}{Numeric value to replace the individual species codes with a complex code shared across species. (default = 3005012)}

\item{by_strata}{Boolean. Should length/age pop'n values be computed at stratum level? (default = FALSE)}

\item{global}{Boolean. Fill in missing length bins with global age-lenth key? (default = FALSE)}

\item{cmplx}{Complex name for saving results. (default = 'bsre')}

\item{region}{Region will create a folder and place results in said folder. (default = 'ai')}

\item{save_interm}{Boolean. Save the intermediate results: resampled age/length comps and realized sample size per iteration? (default = FALSE)}

\item{save_stats}{Boolean. Save other statistics: base age/length comps without resampling, mean length-at-age, bootstrap bias? (default = FALSE)}

\item{save}{Name to attach to and identify output files.}
}
\value{
Dataframe of input sample size by year, complex (designated with cmplx_code), sex (0 - combined sex 
with age/length data combined prior to expansion, 1 - males, 2 - females, 3 - unsexed, 12 - female/male
compositions that sum to one across both sexes combined, 4 - combined sex after summing sex-specific age/length
composition after expansion; all with short description IDs in sex_desc column) for age composition (output saved with
'iss_age' in file name) and length composition (output saved as 'iss_ln' in file name). For comparison, nominal 
sample size ('nss' - the number of age/length samples actually taken) and the number of sampled hauls for 
age/length ('nhls') are included. AI complex output will be denoted with whatever is defined in cmplx argument
 within the file name. Will also produce other dataframes if desired (see save_intern and save_stats argument descriptions).
}
\description{
Bootstrap data sources to replicate bottom trawl survey age and length composition
for computation of input sample size tailored to AI stock complexes (where length composition
is expanded for each individual species in the complex, but age-length specimen data is combined 
across species in the complex prior to age composition expansion)
}
