% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_iss.R
\name{srvy_iss_caal}
\alias{srvy_iss_caal}
\title{replicate survey input sample size function for production run for conditional age-at-length data}
\usage{
srvy_iss_caal(
  iters = 1,
  specimen_data,
  cpue_data,
  r_t,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  region = NULL,
  save_interm = FALSE,
  save_stats = FALSE,
  save
)
}
\arguments{
\item{iters}{number of iterations (500 recommended)}

\item{specimen_data}{input dataframe}

\item{cpue_data}{input dataframe}

\item{r_t}{input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{bin}{bin size for length data}

\item{boot_hauls}{resample hauls w/replacement (default = FALSE)}

\item{boot_ages}{resample ages w/replacement (default = FALSE)}

\item{al_var}{include age-length variability (default = FALSE)}

\item{al_var_ann}{resample age-length annually or pooled across years}

\item{age_err}{include ageing error (default = FALSE)}

\item{region}{region will create a folder and place results in said folder}

\item{save_interm}{save the intermediate results: original comps, resampled comps (default = FALSE)}

\item{save_stats}{save other statistics: base age/length comps, mean length at age, bootstrap bias (default = FALSE)}

\item{save}{name to save output}
}
\description{
replicate survey input sample size function for production run for conditional age-at-length data
}
