% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_iss_exp.R
\name{srvy_iss_exp}
\alias{srvy_iss_exp}
\title{replicate survey input sample size function for experimental runs}
\usage{
srvy_iss_exp(
  iters = 1,
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  r_t,
  yrs = NULL,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  age_err = FALSE,
  region = NULL,
  save_orig = FALSE,
  save_comps = FALSE,
  save_ess = FALSE,
  match_orig = FALSE,
  srvy_type = NULL
)
}
\arguments{
\item{iters}{number of iterations (500 recommended)}

\item{lfreq_data}{input dataframe}

\item{specimen_data}{input dataframe}

\item{cpue_data}{input dataframe}

\item{strata_data}{input dataframe}

\item{r_t}{input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot_hauls}{resample hauls w/replacement (default = FALSE)}

\item{boot_lengths}{resample lengths w/replacement (default = FALSE)}

\item{boot_ages}{resample ages w/replacement (default = FALSE)}

\item{al_var}{include age-length variability (default = FALSE)}

\item{age_err}{include ageing error (default = FALSE)}

\item{region}{region will create a folder and place results in said folder}

\item{save_orig}{save the original comps (default = FALSE)}

\item{save_comps}{save the resampled comps (default = FALSE)}

\item{save_ess}{save the iterated effective sample sizes (default = FALSE)}

\item{match_orig}{match the computed values to gap output (default = FALSE)}

\item{srvy_type}{only for bering sea survey, denotes whether it's the shelf or slope survey (default = NULL)}
}
\description{
replicate survey input sample size function for experimental runs
}
