% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyISS-package.R
\docType{package}
\name{surveyISS-package}
\alias{surveyISS}
\alias{surveyISS-package}
\title{surveyISS: Survey composition input sample size}
\description{
Produce bootstrap estimates of input sample sizes for use in statistical catch at age fishery assessments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BenWilliams-NOAA/surveyISS/}
  \item \url{https://benwilliams-noaa.github.io/surveyISS/}
  \item Report bugs at \url{https://github.com/BenWilliams-NOAA/surveyISS/issues}
}

}
\author{
\strong{Maintainer}: Ben Williams \email{ben.williams@noaa.gov} (\href{https://orcid.org/0000-0001-7295-2076}{ORCID})

Authors:
\itemize{
  \item Pete Hulson \email{pete.hulson@noaa.gov}
}

}
\keyword{internal}
