% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_error.R
\name{age_error}
\alias{age_error}
\title{Resample reader/tester data to implement ageing error.}
\usage{
age_error(age_dat, r_t, nonpar = FALSE)
}
\arguments{
\item{age_dat}{age-length specimen input dataframe}

\item{r_t}{reader/tester agreement data}

\item{nonpar}{Boolean. Use either sampling w/replacement (nonparameteric) or multinomial (parameteric) for reader-tester data (default = FALSE)}
}
\description{
Resample reader/tester data to implement ageing error.
}
