% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_error.R
\name{age_error}
\alias{age_error}
\title{resample reader/tester data to implement ageing error}
\usage{
age_error(age_dat, r_t)
}
\arguments{
\item{age_dat}{age specimen data}

\item{r_t}{reader/tester data}
}
\description{
resample reader/tester data to implement ageing error
}
