% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_variab.R
\name{al_variab}
\alias{al_variab}
\title{Resample age-length data to implement growth variability in age-length key.}
\usage{
al_variab(age_dat, annual = FALSE)
}
\arguments{
\item{age_dat}{age-length specimen input dataframe}

\item{annual}{resample age-length data by years or pooled across years}
}
\description{
Resample age-length data to implement growth variability in age-length key.
}
