% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apop.R
\name{apop_caal}
\alias{apop_caal}
\title{Compute conditional age-at-length.}
\usage{
apop_caal(agedat)
}
\arguments{
\item{agedat}{age-length specimen input dataframe}
}
\value{
dataframe of proportions-at-age by length
}
\description{
Compute conditional age-at-length.
}
