% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{bias_length}
\alias{bias_length}
\title{calculate bias in bootstrapped length comp}
\usage{
bias_length(r_length, ogl)
}
\arguments{
\item{r_length}{iterated length pop'n}

\item{ogl}{original length comps without any sampling}
}
\description{
calculate bias in bootstrapped length comp
}
