% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{comp_stats_caal}
\alias{comp_stats_caal}
\title{wrapper function to compute statistics of bootstrap resampling for conditional age-at-length}
\usage{
comp_stats_caal(r_caal, ogcaal, specimen_data)
}
\arguments{
\item{r_caal}{list of conditional age-at-lengthdata}

\item{ogcaal}{original conditional age-at-length data (single list)}

\item{specimen_data}{input dataframe}

\item{lfreq_data}{input dataframe}
}
\description{
wrapper function to compute statistics of bootstrap resampling for conditional age-at-length
}
