% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{grwth_stats}
\alias{grwth_stats}
\title{calculate mean length-at-age and sd}
\usage{
grwth_stats(r_age, oga)
}
\arguments{
\item{r_age}{iterated age pop'n}

\item{oga}{original age comps without any sampling}
}
\description{
calculate mean length-at-age and sd
}
