% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{iss_age}
\alias{iss_age}
\title{calculate age comp input sample size and add nominal sample size and number of sampled hauls to output}
\usage{
iss_age(rss_age, specimen_data, survey_region = NULL)
}
\arguments{
\item{rss_age}{iterated realized sample size}

\item{specimen_data}{input dataframe}

\item{survey_region}{overall region, e.g., goa or ai (default = NULL)}
}
\description{
calculate age comp input sample size and add nominal sample size and number of sampled hauls to output
}
