% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{iss_caal}
\alias{iss_caal}
\title{calculate conditional-age-at-length input sample size and add nominal sample size and number of sampled hauls to output}
\usage{
iss_caal(rss_caal, specimen_data)
}
\arguments{
\item{rss_caal}{iterated realized sample size}

\item{specimen_data}{input dataframe}
}
\description{
calculate conditional-age-at-length input sample size and add nominal sample size and number of sampled hauls to output
}
