% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpop.R
\name{lpop_gap}
\alias{lpop_gap}
\title{population at length following gapindex package}
\usage{
lpop_gap(lfreq_un, cpue, by_strata = FALSE)
}
\arguments{
\item{lfreq_un}{uncounted length frequency data}

\item{cpue}{cpue data}

\item{by_strata}{are the results to be given by strata, or summed to region level (default = FALSE)}
}
\description{
population at length following gapindex package
}
