% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\title{query data to run surveyISS}
\usage{
query_data(survey, region, species, yrs = NULL)
}
\arguments{
\item{survey}{survey number for gap_products survey_definition_id (ai = 52, goa = 47, ebs = 98, nbs = 143, ebs slope = 78)}

\item{region}{region description for output data file}

\item{species}{species_codes e.g., c(10110, 21740)}

\item{yrs}{minimum year to consider (default: NULL)}
}
\description{
query data to run surveyISS
}
