% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_stats.R
\name{rss_age}
\alias{rss_age}
\title{calculate realized sample size for age comps}
\usage{
rss_age(sim_data, og_data)
}
\arguments{
\item{sim_data}{list of abundance by age data}

\item{og_data}{original abundance by age data (single list)}
}
\description{
calculate realized sample size for age comps
}
