% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_comps.R
\name{srvy_comps_ai_cmplx}
\alias{srvy_comps_ai_cmplx}
\title{Resample data and compute length and age population expansion}
\usage{
srvy_comps_ai_cmplx(
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  r_t,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  cmplx_code = NULL,
  use_gapindex = TRUE,
  by_strata = FALSE,
  global = FALSE
)
}
\arguments{
\item{lfreq_data}{length frequency input dataframe}

\item{specimen_data}{age-length specimen input dataframe}

\item{cpue_data}{catch-per-unit effort input dataframe}

\item{strata_data}{strata id and area size input dataframe}

\item{r_t}{age reader-tester input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{bin}{bin size (default = 1 cm)}

\item{boot_hauls}{Boolean. Resample hauls w/replacement? (default = FALSE)}

\item{boot_lengths}{Boolean. Resample length frequency w/replacement? (default = FALSE)}

\item{boot_ages}{Boolean. Resample ages w/replacement? (default = FALSE)}

\item{al_var}{Boolean. Include age-length variability in resampled age data? (default = FALSE)}

\item{al_var_ann}{Boolean. Resample age-length variability annually or pooled across years? (default = FALSE)}

\item{age_err}{Boolean. Include ageing error in resampled age data? (default = FALSE)}

\item{cmplx_code}{Numeric value to replace the individual species codes with a complex code shared across species. (default = 3005012)}

\item{use_gapindex}{Boolean. Use functions derived from gapindex package? (default = TRUE)}

\item{by_strata}{Boolean. Should length/age pop'n values be computed at stratum level in gap fcns? (default = FALSE)}

\item{global}{Boolean. Fill in missing length bins with global age-lenth key in gap fcns? (default = FALSE)}
}
\value{
List with dataframes of population numbers at length (.lpop) and age (.apop).
}
\description{
Function to resample data and compute length and age expansion following Fig. 1 in Hulson and Williams 2024. 
Customized for ai complexes (i.e., blackspotted-rougheye rockfish) where population at length are expanded at species 
level, but age specimen data are combined and population at age are expanded at complex level
}
