% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_comps_caal.R
\name{srvy_comps_caal}
\alias{srvy_comps_caal}
\title{primary survey age/length pop'n numbers function}
\usage{
srvy_comps_caal(
  specimen_data,
  cpue_data,
  r_t,
  yrs,
  boot_hauls,
  boot_ages,
  sex_spec,
  al_var,
  al_var_ann,
  age_err
)
}
\arguments{
\item{specimen_data}{age-length specimen data}

\item{cpue_data}{abundance by length data}

\item{r_t}{reader/tester ageing data}

\item{yrs}{age filter returns years >= (default = NULL)}

\item{boot_hauls}{switch for resampling hauls (default = FALSE)}

\item{boot_ages}{switch for resampling ages (default = FALSE)}

\item{sex_spec}{determine whether to do sex specific or total comps (default = TRUE)}

\item{al_var}{switch for including age-length variability (default = FALSE)}

\item{al_var_ann}{resample age-length annually or pooled across years}

\item{age_err}{switch for including ageing error (default = FALSE)}
}
\description{
primary survey age/length pop'n numbers function
}
