% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_comps.R
\name{srvy_comps_caal}
\alias{srvy_comps_caal}
\title{Resample data and compute conditional age-at-length}
\usage{
srvy_comps_caal(
  specimen_data,
  cpue_data,
  r_t,
  yrs = NULL,
  bin = 1,
  boot_hauls = FALSE,
  boot_ages = FALSE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  age_samples = NULL,
  plus_len = NULL,
  plus_age = NULL
)
}
\arguments{
\item{specimen_data}{age-length specimen input dataframe}

\item{cpue_data}{catch-per-unit effort input dataframe}

\item{r_t}{age reader-tester input dataframe, included in surveyISS package data}

\item{yrs}{any year filter >= (default = NULL)}

\item{bin}{bin size (default = 1 cm), also can use custom length bins following ss3 bin convention}

\item{boot_hauls}{Boolean. Resample hauls w/replacement? (default = FALSE)}

\item{boot_ages}{Boolean. Resample ages w/replacement? (default = FALSE)}

\item{al_var}{Boolean. Include age-length variability in resampled age data? (default = FALSE)}

\item{al_var_ann}{Boolean. Resample age-length variability annually or pooled across years? (default = FALSE)}

\item{age_err}{Boolean. Include ageing error in resampled age data? (default = FALSE)}

\item{age_samples}{If set at a value, tests reductions (and increases) in survey-level number of ages collected. To test, set at a proportion of ages collected, i.e., 0.8 or 1.2 (default = NULL)}

\item{plus_len}{If set at a value, computes length expansion with a plus-length group (default = FALSE)}

\item{plus_age}{If set at a value, computes age expansion with a plus-age group (default = FALSE)}
}
\value{
List with dataframe of conditional age-at-length (.caal).
}
\description{
This function resamples cpue and age specimen data, adds ageing error and growth variability,
and then computes conditional age-at-length.
}
