% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvy_iss_caal.R
\name{srvy_iss_caal}
\alias{srvy_iss_caal}
\title{replicate survey input sample size function for production run}
\usage{
srvy_iss_caal(
  iters = 1,
  specimen_data,
  cpue_data,
  r_t,
  yrs = NULL,
  boot_hauls = FALSE,
  boot_ages = FALSE,
  sex_spec = TRUE,
  al_var = FALSE,
  al_var_ann = FALSE,
  age_err = FALSE,
  region = NULL,
  save_interm = FALSE,
  match_orig = FALSE,
  srvy_type = NULL,
  save
)
}
\arguments{
\item{iters}{number of iterations (500 recommended)}

\item{specimen_data}{input dataframe}

\item{cpue_data}{input dataframe}

\item{r_t}{input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot_hauls}{resample hauls w/replacement (default = FALSE)}

\item{boot_ages}{resample ages w/replacement (default = FALSE)}

\item{sex_spec}{determine whether to do sex specific or total comps (default = TRUE)}

\item{al_var}{include age-length variability (default = FALSE)}

\item{al_var_ann}{resample age-length annually or pooled across years}

\item{age_err}{include ageing error (default = FALSE)}

\item{region}{region will create a folder and place results in said folder}

\item{save_interm}{save the intermediate results: original comps, resampled comps (default = FALSE)}

\item{match_orig}{match the computed values to gap output (default = FALSE)}

\item{srvy_type}{only for bering sea survey, denotes whether it's the shelf or slope survey (default = NULL)}

\item{save}{name to save output}
}
\description{
replicate survey input sample size function for production run
}
