% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyISS-package.R
\docType{package}
\name{surveyISS-package}
\alias{surveyISS}
\alias{surveyISS-package}
\title{surveyISS: Survey age composition input sample sizes with ageing error}
\description{
Produce bootstrap estimates of input sample sizes for use in statistical catch at age fishery assessments
}
\author{
\strong{Maintainer}: Ben Williams \email{ben.williams@noaa.gov} (\href{https://orcid.org/000-0001-7295-2076}{ORCID})

Authors:
\itemize{
  \item Pete Hulson \email{pete.hulson@noaa.gov}
}

}
\keyword{internal}
